% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_column_for_noro_model_type.R
\name{add_column_for_noro_model_type}
\alias{add_column_for_noro_model_type}
\title{Add a column to distinguish between nowcasts produced by baselinenowcast and
those produced in Mellor et al.}
\usage{
add_column_for_noro_model_type(df)
}
\arguments{
\item{df}{Data.frame containing scores, coverage, or quantiled nowcasts
for each of the three baselinenowcast model configurations}
}
\value{
Data.frame with a new column for model type.
}
\description{
Add a column to distinguish between nowcasts produced by baselinenowcast and
those produced in Mellor et al.
}
