% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_baseline_nowcast_pipeline.R
\name{baselinenowcast_pipeline}
\alias{baselinenowcast_pipeline}
\title{Run \code{baselinenowcast} pipeline to generate an individual instance of a
probabilistic nowcast}
\usage{
baselinenowcast_pipeline(
  long_df,
  nowcast_date,
  max_delay,
  n_history_delay,
  n_history_uncertainty,
  n_draws = 1000,
  long_df_for_borrow = NULL,
  borrow_delay = FALSE,
  borrow_uncertainty = FALSE
)
}
\arguments{
\item{long_df}{Data.frame containing the following columns:
\code{reference_date}, \code{report_date}, and \code{count} indicating the cases on
each reference and report date. This might be filtered such that only
reference dates from one day of the week are included}

\item{nowcast_date}{String indicating the date we want to generate a nowcast.}

\item{max_delay}{Integer indicating the maximum delay we want to consider in
the \code{baselinenowcast} model.}

\item{n_history_delay}{Integer indicating the number of reference time
observations to use to estimate the delay.}

\item{n_history_uncertainty}{Integer indicating the number of reporting
triangles to use to generate the estimate of the dispersion parameters}

\item{n_draws}{Integer indicating the number of draws from the observation
model to use to generate the expected observed nowcasts. Default is
\code{1000}.}

\item{long_df_for_borrow}{Data.frame containing the same columns as
\code{long_df} but from the strata to borrow delay or uncertainty from. Default
is \code{NULL} corresponding to no borrowing base case.}

\item{borrow_delay}{Boolean indicating whether to use the
\code{long_df_for_borrow} to estimate the delay distribution, default is
\code{FALSE}.}

\item{borrow_uncertainty}{Boolean indicating whether to use the
\code{long_df_for_borrow} to estimate the dispersion parameters, default is
\code{FALSE}.}
}
\value{
\code{ind_nowcast} A dataframe of the expected observed total counts
for each reference date up until the nowcast date
}
\description{
This function ingests a long form dataframe containing cases by reference
date and report date, the data we want to nowcast from, and the
\code{baselinenowcast} model specifications and returns a probabilistic
nowcast of the expected total cases by reference date up until the
specified nowcast date. Note this is not being used.
}
