% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{derive_model_variation}
\alias{derive_model_variation}
\title{Derive model variations}
\usage{
derive_model_variation(df)
}
\arguments{
\item{df}{dataframe with columns of metadata on covid model permutation
variations}
}
\value{
dataframe with two new columns for model_variation and
model_variation_string
}
\description{
Derive model variations
}
