% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_uncertainty.R
\name{estimate_uncertainty}
\alias{estimate_uncertainty}
\title{Estimate uncertainty using baselinenowcast functions}
\usage{
estimate_uncertainty(
  triangle_for_uncertainty,
  n_history_uncertainty,
  n_history_delay,
  fun_to_aggregate = sum,
  k = 7
)
}
\arguments{
\item{triangle_for_uncertainty}{Matrix of reporting triangle to use for
uncertainty estimation.}

\item{n_history_uncertainty}{Integer indicating number of retrospective
nowcast dates to use for uncertainty estimation}

\item{n_history_delay}{Integer indicating number of reference times to use
for delay estimation}

\item{fun_to_aggregate}{Function for aggregating the target data, default
is \code{sum}.}

\item{k}{Integer indicating window size for the aggregate function to
operate over, default is \code{7}.}
}
\value{
Vector of dispersion parameters
}
\description{
Estimate uncertainty using baselinenowcast functions
}
