% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{get_bar_chart_sum_scores_noro}
\alias{get_bar_chart_sum_scores_noro}
\title{Plot stacked bar chart of scores by model type colored by model and
with alpha by over/under prediction and dispersion}
\usage{
get_bar_chart_sum_scores_noro(scores)
}
\arguments{
\item{scores}{Dataframe containing scores across
nowcast dates and reference dates, stratified by model and and model type.}
}
\value{
ggplot object
}
\description{
Plot stacked bar chart of scores by model type colored by model and
with alpha by over/under prediction and dispersion
}
