% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_data.R
\name{get_covid_data}
\alias{get_covid_data}
\title{Get COVID data from German Nowcast Hub}
\usage{
get_covid_data(url, loc_to_subset = "DE")
}
\arguments{
\item{url}{String of url for raw data}

\item{loc_to_subset}{String indicating the location to subset the data to,
default is "DE" which will get the national data for all age groups.}
}
\value{
Data.frame of counts of cases by reference date and report date for
only Germany and
}
\description{
Get COVID data from German Nowcast Hub
}
