% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eval_data_from_long_df.R
\name{get_eval_data_from_long_df}
\alias{get_eval_data_from_long_df}
\title{Get aggregated data as of a certain date from long tidy dataframe}
\usage{
get_eval_data_from_long_df(long_df, as_of_date, max_delay = Inf)
}
\arguments{
\item{long_df}{Dataframe of the latest data by reference and report date}

\item{as_of_date}{String indicating the as of date}

\item{max_delay}{Integer indicating the maximum delays. Default is \code{Inf},
which means we would not restrict to only delays less than the
maximum delay.}
}
\value{
Data.frame summarised by reference time as of the as of date
}
\description{
Get aggregated data as of a certain date from long tidy dataframe
}
