% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kit_nowcast_url.R
\name{get_kit_nowcast_url}
\alias{get_kit_nowcast_url}
\title{Get the url to load in the KIT nowcast data}
\usage{
get_kit_nowcast_url(prefix, nowcast_date)
}
\arguments{
\item{prefix}{Character string indicating the github url up until the date
appears.}

\item{nowcast_date}{Character string indicating the nowcast date.}
}
\value{
Character string indicating the url for the nowcast on that nowcast
date.
}
\description{
Get the url to load in the KIT nowcast data
}
