% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noro_delay_outputs.R
\name{get_noro_delay_outputs}
\alias{get_noro_delay_outputs}
\title{Get norovirus delay outputs}
\usage{
get_noro_delay_outputs(
  noro_df,
  nowcast_date,
  max_delay,
  filter_ref_dates,
  n_history_delay
)
}
\arguments{
\item{noro_df}{Dataframe of the latest data by reference and
report date}

\item{nowcast_date}{Character string in YYYY-MM-DD format indicating the
date of the nowcast.}

\item{max_delay}{Integer indicating the maximum delay.}

\item{filter_ref_dates}{Boolean indicating whether or not to make separate
delay estimates by weekday.}

\item{n_history_delay}{Integer indicating number of reference times used for
delay estimate.}
}
\value{
Data.frame of delays with metadata
}
\description{
Get norovirus delay outputs
}
