% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nowcast_quantiles.R
\name{get_nowcast_quantiles}
\alias{get_nowcast_quantiles}
\title{Get nowcast quantiles from point nowcast, dispersion, and other metadata}
\usage{
get_nowcast_quantiles(
  point_nowcast_matrix,
  reporting_triangle,
  dispersion,
  draws,
  nowcast_date,
  date_df,
  data_as_of_df,
  eval_data,
  days_to_eval,
  model,
  n_history_delay,
  n_history_uncertainty,
  borrow,
  partial_rep_tri,
  age_group,
  quantiles,
  fun_to_aggregate = sum,
  k = 7
)
}
\arguments{
\item{point_nowcast_matrix}{Matrix containing the point estimate of the
nowcast elements with both observations and predictions.}

\item{reporting_triangle}{Matrix original reporting triangle
(observations only).}

\item{dispersion}{Vector of dispersion parameters.}

\item{draws}{Integer indicating number of draws to sample.}

\item{nowcast_date}{Character string in YYYY-MM-DD format indicating the
date of the nowcast.}

\item{date_df}{Dataframe of reference dates and time indices}

\item{data_as_of_df}{Dataframe containing the total counts across delays
for each reference time.}

\item{eval_data}{Dataframe containing the final counts to evaluate against
(should be transformed to the target already).}

\item{days_to_eval}{Integer indicating the number of days before the nowcast
date to evaluate against.}

\item{model}{Character string indicating which model is being used}

\item{n_history_delay}{Integer indicating number of reference times used for
delay estimate.}

\item{n_history_uncertainty}{Integer indicating number of retrospective
nowcast datasets to use for uncertainty estimation}

\item{borrow}{Boolean indicating whether or not the delay and dispersion
is borrowed from another strata.}

\item{partial_rep_tri}{Boolean indicating whether or not a partially
complete reporting triangle is used or the latest complete triangle is
used.}

\item{age_group}{Character string indicating the age group.}

\item{quantiles}{Vector of quantiles to generate.}

\item{fun_to_aggregate}{Function to apply across the \code{k} reference dates,
default is \code{sum}.}

\item{k}{Width of reference dates to apply transform to, default is \code{1}}
}
\value{
scoringutils forecast quantile object
}
\description{
Get nowcast quantiles from point nowcast, dispersion, and other metadata
}
