% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{get_plot_bar_chart_scores_mp}
\alias{get_plot_bar_chart_scores_mp}
\title{Get horizontal bar chart of summarised scores by model permutation and score
component}
\usage{
get_plot_bar_chart_scores_mp(scores, strata = "age groups")
}
\arguments{
\item{scores}{Data.frame of model permutations for all nowcast dates,
horizons, and age groups.}

\item{strata}{Character string indicating whether to summarise across the
different age strata (\code{"age groups"}) or across the nation as a whole
(\code{"national"}). Default is \code{"age groups"}.}
}
\value{
ggplot object
}
\description{
Get horizontal bar chart of summarised scores by model permutation and score
component
}
