% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_hub_validation.R
\name{get_plot_bar_chart_sum_scores}
\alias{get_plot_bar_chart_sum_scores}
\title{Get horizontal bar chart of summarised scores by model and score component}
\usage{
get_plot_bar_chart_sum_scores(joined_scores, strata = "age groups")
}
\arguments{
\item{joined_scores}{Data.frame of a single model specification of
baselinenowcast + a comparison model, with both models spanning the same
nowcast dates and age groups}

\item{strata}{Character string indicating whether to summarise across the
different age strata (\code{"age groups"}) or across the nation as a whole
(\code{"national"}). Default is \code{"age groups"}.}
}
\value{
ggplot object
}
\description{
Get horizontal bar chart of summarised scores by model and score component
}
