% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{get_plot_coverage_by_model_noro}
\alias{get_plot_coverage_by_model_noro}
\title{Get a plot of coverage by model}
\usage{
get_plot_coverage_by_model_noro(
  all_coverage,
  intervals = c(50, 90),
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp"),
  save = TRUE
)
}
\arguments{
\item{all_coverage}{Dataframe of the combined individual level coverage for
days and intervals.}

\item{intervals}{Vector of integers to plot coverage of.}
}
\value{
bar chart
}
\description{
Get a plot of coverage by model
}
