% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_eda.R
\name{get_plot_data_as_of}
\alias{get_plot_data_as_of}
\title{Get a plot of the data as of different dates}
\usage{
get_plot_data_as_of(final_df, as_of_dates, pathogen = "")
}
\arguments{
\item{final_df}{Dataframe of the latest data by reference and report date}

\item{as_of_dates}{Vector of character strings of the dates you wish to plot}

\item{pathogen}{Character sting of the pathogen being plotted}
}
\value{
ggplot object
}
\description{
Get a plot of the data as of different dates
}
