% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_eda.R
\name{get_plot_ind_nowcast_draws}
\alias{get_plot_ind_nowcast_draws}
\title{Get a plot of the draws of an individual nowcast}
\usage{
get_plot_ind_nowcast_draws(
  nowcast_draws,
  nowcast_target = "7-day rolling sum COVID-19 admissions"
)
}
\arguments{
\item{nowcast_draws}{Dataframe of the draws of the nowcast with data}

\item{nowcast_target}{Character string indicating the quantity being
nowcasted}
}
\value{
ggplot object
}
\description{
Get a plot of the draws of an individual nowcast
}
