% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_hub_validation.R
\name{get_plot_mean_delay_over_time}
\alias{get_plot_mean_delay_over_time}
\title{Plot the mean delay over time colored by age group}
\usage{
get_plot_mean_delay_over_time(delays_over_time)
}
\arguments{
\item{delays_over_time}{Dataframe containing the average delay by nowcast
date and age group, including the national data}
}
\value{
ggplot object
}
\description{
Plot the mean delay over time colored by age group
}
