% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{get_plot_mean_delay_over_time_by_weekday}
\alias{get_plot_mean_delay_over_time_by_weekday}
\title{Get plot mean delay over time by weekday}
\usage{
get_plot_mean_delay_over_time_by_weekday(delay_dfs)
}
\arguments{
\item{delay_dfs}{Data.frame of delay estimates both separately by weekday
and jointly, indexed by the weekday of the reference date.}
}
\value{
ggplot object
}
\description{
Get plot mean delay over time by weekday
}
