% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_eda.R
\name{get_plot_mult_nowcasts}
\alias{get_plot_mult_nowcasts}
\title{Get a plot of the nowcasts string together for all dates or a set of dates}
\usage{
get_plot_mult_nowcasts(
  all_nowcasts,
  final_summed_data,
  nowcast_dates_to_plot = NULL,
  pathogen = "",
  title = "",
  facet = FALSE
)
}
\arguments{
\item{all_nowcasts}{Dataframe of the quantiled nowcasts (in wide format)}

\item{final_summed_data}{Dtaaframe of the sum of the counts by reference
date as of the final reference data + the eval time frame.}

\item{nowcast_dates_to_plot}{Vector of character strings of the dates you
wish to plot, default is \code{NULL} which will plot all of them}

\item{pathogen}{Character sting of the pathogen being plotted}

\item{title}{Character string indicating the title}

\item{facet}{Boolean indicating whether or not to make separate facets
of each model}
}
\value{
ggplot object
}
\description{
Get a plot of the nowcasts string together for all dates or a set of dates
}
