% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_hub_validation.R
\name{get_plot_nowcast_illustration}
\alias{get_plot_nowcast_illustration}
\title{Get a plot illustrating nowcasts at certain dates}
\usage{
get_plot_nowcast_illustration(
  combined_nowcasts,
  max_horizon = 28,
  nowcast_dates_to_plot = NULL,
  age_group_to_plot = "00+",
  facet = FALSE
)
}
\arguments{
\item{combined_nowcasts}{Dataframe of the combined quantiles across
horizons and nowcast dates}

\item{max_horizon}{Integer indicating the maximum horizon to plot.}

\item{nowcast_dates_to_plot}{Vector of character strings of the dates you
wish to plot, default is \code{NULL} which will plot all of them}

\item{age_group_to_plot}{Character string indicating which age group to
show in the plot, default is "00+" for all age groups.}

\item{facet}{Boolean indicating whether or not to make separate facets
of each model}
}
\value{
ggplot object
}
\description{
Hub validation figure A
}
