% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{get_plot_nowcast_perms_over_time}
\alias{get_plot_nowcast_perms_over_time}
\title{Get a plot of the nowcasts over time faceted by model variation, colored
by the type of permutation}
\usage{
get_plot_nowcast_perms_over_time(
  combined_nowcasts,
  horizon_to_plot,
  age_group_to_plot = "00+"
)
}
\arguments{
\item{combined_nowcasts}{Dataframe of the combined quantiles across
horizons and nowcast dates for all the model permutations.}

\item{horizon_to_plot}{Integer indicating which horizon to plot}

\item{age_group_to_plot}{Character string indicating which age group to
show in the plot, default is "00+" for all age groups.}
}
\value{
ggplot object
}
\description{
Hub validation figure A
}
