% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_hub_validation.R
\name{get_plot_nowcasts_over_time}
\alias{get_plot_nowcasts_over_time}
\title{Get a plot of the nowcasts over time for both KIT and baselinenowcast}
\usage{
get_plot_nowcasts_over_time(combined_nowcasts, horizon_to_plot, facet = FALSE)
}
\arguments{
\item{combined_nowcasts}{Dataframe of the combined quantiles across
horizons and nowcast dates}

\item{horizon_to_plot}{Integer indicating which horizon to plot}

\item{facet}{Boolean indicating whether or not to make separate facets
of each model}
}
\value{
ggplot object
}
\description{
Hub validation figure A
}
