% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{get_plot_nowcasts_over_time_mp}
\alias{get_plot_nowcasts_over_time_mp}
\title{Get a plot of the nowcasts over time faceted by model variation, colored
by the type of permutation}
\usage{
get_plot_nowcasts_over_time_mp(
  combined_nowcasts,
  horizon_to_plot,
  age_group_to_plot = "00+",
  permutation_grouping = NULL,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp"),
  save = TRUE
)
}
\arguments{
\item{combined_nowcasts}{Dataframe of the combined quantiles across
horizons and nowcast dates for all the model permutations.}

\item{horizon_to_plot}{Integer indicating which horizon to plot}

\item{age_group_to_plot}{Character string indicating which age group to
show in the plot. Default is \code{"00+"}for all age groups.}

\item{permutation_grouping}{Character string indicating which model
permutation grouping to plot. Default is \code{NULL} for all of them.}

\item{fig_file_name}{Character string indicating name of the figure to be
saved as the file name. Default is \code{NULL}.}

\item{fig_file_dir}{Path to save figure. Default is
\code{file.path("output", "figs", "supp")}.}

\item{save}{Boolean indicating whether or not to save the figure to disk.
Default is \code{TRUE}.}
}
\value{
ggplot object
}
\description{
Hub validation figure A
}
