% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_hub_validation.R
\name{get_plot_of_delay_cdf_by_age}
\alias{get_plot_of_delay_cdf_by_age}
\title{Plot the delay cdf by age group and nationally}
\usage{
get_plot_of_delay_cdf_by_age(avg_delays_by_age)
}
\arguments{
\item{avg_delays_by_age}{Dataframe containing the average delay cdf across
nowcast dates for each age group}
}
\value{
ggplot object
}
\description{
Plot the delay cdf by age group and nationally
}
