% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{get_plot_rel_decomposed_wis}
\alias{get_plot_rel_decomposed_wis}
\title{Get a plot of relative decomposed WIS by age group}
\usage{
get_plot_rel_decomposed_wis(scores, facet = FALSE)
}
\arguments{
\item{scores}{Dataframe of all the scores by individual reference and
nowcast dates and model permutation}

\item{facet}{Boolean indicating whether or not to create separate facets for
each WIS component. Default is \code{FALSE}}
}
\value{
ggplot object
}
\description{
Get a plot of relative decomposed WIS by age group
}
