% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{get_plot_rel_delay_t_by_wday}
\alias{get_plot_rel_delay_t_by_wday}
\title{Get plot relative mean delay over time by weekday}
\usage{
get_plot_rel_delay_t_by_wday(
  delay_dfs,
  n_history_delay_filter = 28,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp"),
  save = TRUE
)
}
\arguments{
\item{delay_dfs}{Data.frame of delay estimates both separately by weekday
and jointly, indexed by the weekday of the reference date.}

\item{n_history_delay_filter}{Integer indicating how much historical data
to use to estimate the mean delay at each nowcast time. Default is \code{14}.}

\item{fig_file_name}{Character string indicating name of the figure to be
saved as the file name. Default is \code{NULL}.}

\item{fig_file_dir}{Path to save figure. Default is
\code{file.path("output", "figs", "supp")}.}

\item{save}{Boolean indicating whether or not to save the figure to disk.
Default is \code{TRUE}.}
}
\value{
ggplot object
}
\description{
Get plot relative mean delay over time by weekday
}
