% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{get_plot_rel_wis_by_horizon_mp}
\alias{get_plot_rel_wis_by_horizon_mp}
\title{Get a plot of relative WIS by horizon for all model permutations}
\usage{
get_plot_rel_wis_by_horizon_mp(scores, strata)
}
\arguments{
\item{scores}{Dataframe of all the scores by individual reference and
nowcast dates and model permutation}

\item{strata}{Character string indicating whether to summarise across the
different age strata (\code{"age groups"}) or across the nation as a whole
(\code{"national"}). Default is \code{"age groups"}.}
}
\value{
ggplot object
}
\description{
Get a plot of relative WIS by horizon for all model permutations
}
