% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{get_plot_rel_wis_by_weekday}
\alias{get_plot_rel_wis_by_weekday}
\title{Get plot relative wis by weekday}
\usage{
get_plot_rel_wis_by_weekday(scores)
}
\arguments{
\item{scores}{Dataframe containing scores across
nowcast dates and reference dates, stratified by model and and model type.}
}
\value{
ggplot object
}
\description{
Get plot relative wis by weekday
}
