% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{get_plot_rel_wis_over_time}
\alias{get_plot_rel_wis_over_time}
\title{Get a line plot of relative WIS by nowcast date}
\usage{
get_plot_rel_wis_over_time(scores)
}
\arguments{
\item{scores}{Dataframe containing scores across
nowcast dates and reference dates, stratified by model and and model type.}
}
\value{
ggplot object
}
\description{
Get a line plot of relative WIS by nowcast date
}
