% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{get_plot_rel_wis_over_time_all}
\alias{get_plot_rel_wis_over_time_all}
\title{Get a line plot of the relative wis over time for each model permutation}
\usage{
get_plot_rel_wis_over_time_all(
  scores,
  strata,
  fig_file_name,
  fig_file_dir = file.path("output", "figs", "supp"),
  save = TRUE
)
}
\arguments{
\item{scores}{Data.frame of model permutations for all nowcast dates,
horizons, and age groups.}

\item{strata}{Character string indicating whether to summarise across the
different age strata (\code{"age groups"}) or across the nation as a whole
(\code{"national"}). Default is \code{"age groups"}.}

\item{fig_file_name}{Character string indicating name of the figure to be
saved as the file name}

\item{fig_file_dir}{Path to save figure. Default is
\code{file.path("output", "figs")}.}

\item{save}{Boolean indicating whether or not to save the figure to disk.
Default is \code{TRUE}.}
}
\value{
ggplot object
}
\description{
Get a line plot of the relative wis over time for each model permutation
}
