% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_hub_validation.R
\name{get_plot_score_by_age_group}
\alias{get_plot_score_by_age_group}
\title{Plot stacked bar chart of scores by age group colored by model and patterned
by over/under prediction and dispersion}
\usage{
get_plot_score_by_age_group(scores_by_age_group)
}
\arguments{
\item{scores_by_age_group}{Dataframe containing scores summarised across
nowcast dates and reference dates, stratified by model and age group}
}
\value{
ggplot object
}
\description{
Plot stacked bar chart of scores by age group colored by model and patterned
by over/under prediction and dispersion
}
