% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{get_plot_wis_by_age_group_mp}
\alias{get_plot_wis_by_age_group_mp}
\title{Get a plot of decomposed WIS by age group for each model permutation}
\usage{
get_plot_wis_by_age_group_mp(scores_by_age_group)
}
\arguments{
\item{scores_by_age_group}{Dataframe of the scores by age group
and model}
}
\value{
ggplot object
}
\description{
Get a plot of decomposed WIS by age group for each model permutation
}
