% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{get_plot_wis_by_weekday}
\alias{get_plot_wis_by_weekday}
\title{Get a plot of decomposed WIS by weekday for each model}
\usage{
get_plot_wis_by_weekday(
  scores,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp"),
  save = TRUE
)
}
\arguments{
\item{scores}{Dataframe of the scores by age group, horizon, and model}

\item{fig_file_name}{Character string indicating name of the figure to be
saved as the file name. Default is \code{NULL}.}

\item{fig_file_dir}{Path to save figure. Default is
\code{file.path("output", "figs", "supp")}.}

\item{save}{Boolean indicating whether or not to save the figure to disk.
Default is \code{TRUE}.}
}
\value{
ggplot object
}
\description{
Get a plot of decomposed WIS by weekday for each model
}
