% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_hub_validation.R
\name{get_plot_wis_over_time}
\alias{get_plot_wis_over_time}
\title{Plot the average WIS by model over time}
\usage{
get_plot_wis_over_time(scores_summarised)
}
\arguments{
\item{scores_summarised}{Dataframe of scores averaged across age groups and
reference times}
}
\value{
ggplot object
}
\description{
Plot the average WIS by model over time
}
