% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rep_tri_from_long_df.R
\name{get_rep_tri_from_long_df}
\alias{get_rep_tri_from_long_df}
\title{Get reporting triangle from long tidy dataframe}
\usage{
get_rep_tri_from_long_df(long_df, nowcast_date, max_delay)
}
\arguments{
\item{long_df}{Dataframe of the latest data by reference and report date}

\item{nowcast_date}{String indicating data to nowcast as of}

\item{max_delay}{Integer indicating the maximum delays}
}
\value{
Dataframe in the form of a reporting triangle
}
\description{
Get reporting triangle from long tidy dataframe
}
