% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sunday_breaks}
\alias{get_sunday_breaks}
\title{Get all the breaks for sundays}
\usage{
get_sunday_breaks(start_date, end_date)
}
\arguments{
\item{start_date}{first date}

\item{end_date}{last date}
}
\value{
vector of sunday dates
}
\description{
Get all the breaks for sundays
}
