% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_triangle.R
\name{get_triangle}
\alias{get_triangle}
\title{Get reporting triangle}
\usage{
get_triangle(long_df, nowcast_date, max_delay, age_group)
}
\arguments{
\item{long_df}{Dataframe with all age groups, reference dates, and report}

\item{nowcast_date}{Character string indicating the date of the nowcast}

\item{max_delay}{Integer indicating the maximum delay}

\item{age_group}{Character string indicating the age group for this triangle}
}
\value{
\code{triangle} Matrix of reporting triangle
}
\description{
Get reporting triangle
}
