% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_hub_validation.R
\name{make_fig_hub_validation}
\alias{make_fig_hub_validation}
\title{Title}
\usage{
make_fig_hub_validation(
  plot_nowcasts_over_time,
  horiz_bar_chart_sum_scores_ag,
  plot_wis_comp_over_time_ag,
  bar_chart_scores_by_age_group,
  plot_mean_delay_over_time_by_age,
  plot_mean_cdf_delay_by_age,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs"),
  save = TRUE
)
}
\arguments{
\item{plot_nowcasts_over_time}{A}

\item{horiz_bar_chart_sum_scores_ag}{B}

\item{plot_wis_comp_over_time_ag}{C}

\item{bar_chart_scores_by_age_group}{D}

\item{plot_mean_delay_over_time_by_age}{E}

\item{plot_mean_cdf_delay_by_age}{F}

\item{fig_file_name}{Character string indicating name of the figure to be
saved as the file name. Default is \code{NULL}.}

\item{fig_file_dir}{Path to save figure. Default is
\code{file.path("output", "figs")}.}

\item{save}{Boolean indicating whether or not to save the figure to disk.
Default is \code{TRUE}.}
}
\value{
ggplot object as a gridded panel
}
\description{
Title
}
