% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{make_fig_model_perms}
\alias{make_fig_model_perms}
\title{Make panel for main model permutation figure}
\usage{
make_fig_model_perms(
  panel_A_nowcasts_over_time,
  bar_chart_wis_by_mp,
  bar_chart_coverage_mp,
  rel_wis_by_horizon_mp,
  rel_decomp_wis_by_age_group,
  fig_file_name,
  fig_file_dir = file.path("output", "figs"),
  save = TRUE
)
}
\arguments{
\item{panel_A_nowcasts_over_time}{A}

\item{bar_chart_wis_by_mp}{B}

\item{bar_chart_coverage_mp}{C}

\item{rel_wis_by_horizon_mp}{D}

\item{rel_decomp_wis_by_age_group}{E}

\item{fig_file_name}{Character string indicating name of the figure to be
saved as the file name}

\item{fig_file_dir}{Path to save figure. Default is
\code{file.path("output", "figs")}.}

\item{save}{Boolean indicating whether or not to save the figure to disk.
Default is \code{TRUE}.}
}
\value{
ggplot object as a gridded panel
}
\description{
Make panel for main model permutation figure
}
