% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{make_fig_noro}
\alias{make_fig_noro}
\title{Make panel for main norovirus figure}
\usage{
make_fig_noro(
  panel_A_noro,
  bar_chart_wis_noro,
  rel_wis_by_weekday,
  wis_by_weekday,
  distrib_mean_delay_weekday,
  plot_mean_delay_t_by_wday,
  plot_cdf_by_weekday,
  fig_file_name,
  fig_file_dir = file.path("output", "figs"),
  save = TRUE
)
}
\arguments{
\item{panel_A_noro}{A}

\item{bar_chart_wis_noro}{B}

\item{rel_wis_by_weekday}{C}

\item{wis_by_weekday}{D}

\item{distrib_mean_delay_weekday}{E}

\item{plot_mean_delay_t_by_wday}{F}

\item{plot_cdf_by_weekday}{G}

\item{fig_file_name}{Character string indicating name of the figure to be
saved as the file name}

\item{fig_file_dir}{Path to save figure. Default is
\code{file.path("output", "figs")}.}

\item{save}{Boolean indicating whether or not to save the figure to disk.
Default is \code{TRUE}.}
}
\value{
ggplot
}
\description{
Make panel for main norovirus figure
}
