% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{make_panel_A_mps_2_ags}
\alias{make_panel_A_mps_2_ags}
\title{Make a patchwork plot for panel A combining facets and underlays with two
age groups side by side}
\usage{
make_panel_A_mps_2_ags(
  plot_nowcasts_t_mp_borrow1,
  rel_wis_over_time_mp_borrow1,
  plot_nowcasts_t_mp_rep_tri1,
  rel_wis_over_time_mp_rep_tri1,
  plot_nowcasts_t_mp_volume1,
  rel_wis_over_time_mp_volume1,
  plot_nowcasts_t_mp_wday1,
  rel_wis_over_time_mp_wday1,
  plot_nowcasts_t_mp_borrow2,
  rel_wis_over_time_mp_borrow2,
  plot_nowcasts_t_mp_rep_tri2,
  rel_wis_over_time_mp_rep_tri2,
  plot_nowcasts_t_mp_volume2,
  rel_wis_over_time_mp_volume2,
  plot_nowcasts_t_mp_wday2,
  rel_wis_over_time_mp_wday2
)
}
\arguments{
\item{plot_nowcasts_t_mp_borrow1}{larger nowcasts over time}

\item{rel_wis_over_time_mp_borrow1}{underlay relative WIS}

\item{plot_nowcasts_t_mp_rep_tri1}{larger nowcasts over time}

\item{rel_wis_over_time_mp_rep_tri1}{underlay relative WIS}

\item{plot_nowcasts_t_mp_volume1}{larger nowcasts over time}

\item{rel_wis_over_time_mp_volume1}{underlay relative WIS}

\item{plot_nowcasts_t_mp_wday1}{larger nowcasts over time}

\item{rel_wis_over_time_mp_wday1}{underlay relative WIS}

\item{plot_nowcasts_t_mp_borrow2}{larger nowcasts over time}

\item{rel_wis_over_time_mp_borrow2}{underlay relative WIS}

\item{plot_nowcasts_t_mp_rep_tri2}{larger nowcasts over time}

\item{rel_wis_over_time_mp_rep_tri2}{underlay relative WIS}

\item{plot_nowcasts_t_mp_volume2}{larger nowcasts over time}

\item{rel_wis_over_time_mp_volume2}{underlay relative WIS}

\item{plot_nowcasts_t_mp_wday2}{larger nowcasts over time}

\item{rel_wis_over_time_mp_wday2}{underlay relative WIS}
}
\value{
patchwork ggplot object
}
\description{
Make a patchwork plot for panel A combining facets and underlays with two
age groups side by side
}
