% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{make_panel_A_noro_nv}
\alias{make_panel_A_noro_nv}
\title{Make panel A norovirus figure}
\usage{
make_panel_A_noro_nv(
  plot_noro_nowcasts_baseline,
  rel_wis_by_week_noro_baseline,
  plot_noro_nowcasts_GAM,
  rel_wis_by_week_noro_GAM,
  plot_noro_nowcasts_enw,
  rel_wis_by_week_noro_enw
)
}
\arguments{
\item{plot_noro_nowcasts_baseline}{nowcasts over time}

\item{rel_wis_by_week_noro_baseline}{relative WIS}

\item{plot_noro_nowcasts_GAM}{nowcasts over time}

\item{rel_wis_by_week_noro_GAM}{relative WIS}

\item{plot_noro_nowcasts_enw}{nowcasts over time}

\item{rel_wis_by_week_noro_enw}{relative WIS}
}
\value{
ggplot
}
\description{
Make panel A norovirus figure
}
