% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Suppress output and messages for code.}
\usage{
quiet(code)
}
\arguments{
\item{code}{Code to run quietly.}
}
\value{
The result of running the code.
}
\description{
Suppress output and messages for code.
}
\examples{
library(cmdstanr)
compile_model("stan/model.stan")
quiet(fit_model("stan/model.stan", simulate_data_discrete()))
out
}
