% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_column_for_noro_model_type.R
\name{replace_Mellor_name_with_ours}
\alias{replace_Mellor_name_with_ours}
\title{Replace model name from Mellor et al with the one we are using for their
outputs}
\usage{
replace_Mellor_name_with_ours(df)
}
\arguments{
\item{df}{Data.frame containing scores, coverage, or quantiled nowcasts
for each of the three baselinenowcast model configurations}
}
\value{
Data.frame with new model names in the \code{model} column.
}
\description{
Replace model name from Mellor et al with the one we are using for their
outputs
}
