% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_baseline_nowcast_pipeline.R
\name{run_baselinenowcast_pipeline}
\alias{run_baselinenowcast_pipeline}
\title{Run \code{baselinenowcast} pipeline}
\usage{
run_baselinenowcast_pipeline(
  long_df,
  nowcast_date,
  max_delay,
  n_history_delay,
  n_history_uncertainty,
  n_draws = 1000,
  reporting_triangle_to_borrow_delay = NULL,
  reporting_triangle_to_borrow_uncertainty = NULL
)
}
\arguments{
\item{long_df}{Data.frame containing the following columns: \code{reference_date},
\code{report_date}, and \code{count} indicating the cases on each reference
and report date.}

\item{nowcast_date}{String indicating the date we want to generate a nowcast.}

\item{max_delay}{Integer indicating the maximum delay we want to consider in
the \code{baselinenowcast} model.}

\item{n_history_delay}{Integer indicating the number of reference time
observations to use to estimate the delay.}

\item{n_history_uncertainty}{Integer indicating the number of reporting
triangles to use to generate the estimate of the dispersion parameters}

\item{n_draws}{Integer indicating the number of draws from the observation
model to use to generate the expected observed nowcasts. Default is
\code{1000}.}

\item{reporting_triangle_to_borrow_delay}{Matrix to be used to estimate the
delay if not the same as the one being nowcasted from, \code{long_df}.}

\item{reporting_triangle_to_borrow_uncertainty}{Matrix to be used to
estimate the dispersion parameters if not the same as the one being
nowcasted from, \code{long_df}.}
}
\value{
\code{summary_nowcast} A dataframe of the expected observed total counts
for each reference date up until the nowcast date
}
\description{
This function ingests a long form dataframe containing cases by reference
date and report date, the data we want to nowcast from, and the
\code{baselinenowcast} model specifications and returns a probabilistic
nowcast of the expected total cases by reference date up until the
specified nowcast date.
}
