% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_covid_nowcast_pipeline.R
\name{run_covid_nowcast_pipeline}
\alias{run_covid_nowcast_pipeline}
\title{Run covid nowcast pipeline}
\usage{
run_covid_nowcast_pipeline(
  long_df_all_strata,
  nowcast_date,
  max_delay,
  age_group_to_nowcast,
  borrow,
  partial_rep_tri,
  weekday_filter,
  n_history_delay,
  n_history_uncertainty,
  eval_timeframe,
  days_to_eval,
  n_draws,
  quantiles,
  plot_quantiles,
  k = 7,
  fun_to_aggregate = sum
)
}
\arguments{
\item{long_df_all_strata}{Dataframe of the latest data by reference and
report date for all age groups}

\item{nowcast_date}{Character string in YYYY-MM-DD format indicating the
date of the nowcast.}

\item{max_delay}{Integer indicating the maximum delay.}

\item{age_group_to_nowcast}{Character string indicating the age group.}

\item{borrow}{Boolean indicating whether or not the delay and dispersion
is borrowed from another strata.}

\item{partial_rep_tri}{Boolean indicating whether or not a partially
complete reporting triangle is used or the latest complete triangle is
used.}

\item{weekday_filter}{Boolean indicating whether or not to estimate delays
and perform nowcasting on individual weekdays and recombine.}

\item{n_history_delay}{Integer indicating number of reference times used for
delay estimate.}

\item{n_history_uncertainty}{Integer indicating number of retrospective
nowcast datasets to use for uncertainty estimation}

\item{eval_timeframe}{Integer indicating the number of days after the
nowcast date to evaluate the nowcasts.}

\item{days_to_eval}{Integer indicating the number of days before the nowcast
date to evaluate against.}

\item{n_draws}{Integer indicating number of draws to sample.}

\item{quantiles}{Vector of quantiles to generate.}

\item{plot_quantiles}{Vector of quantiles for plotting.}

\item{k}{Width of reference dates to apply transform to, default is \code{1}}

\item{fun_to_aggregate}{Function to apply across the \code{k} reference dates,
default is \code{sum}.}
}
\value{
scoringutils forecast quantile object
}
\description{
Run covid nowcast pipeline
}
