% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mellor_et_al_outputs.R
\name{get_mellor_et_al_outputs}
\alias{get_mellor_et_al_outputs}
\title{Placeholder function for extracting the outputs from the Norovirus paper}
\usage{
get_mellor_et_al_outputs(
  data_from_bnc,
  models = c("epinowcast", "GAM", "baseline Mellor et al")
)
}
\arguments{
\item{data_from_bnc}{Data.frame of the output from baselinenowcast}

\item{models}{Vector of character strings indicating which models from
Mellor et al to include. Default is
\code{c("epinowcast", "GAM", "baseline Mellor et al")}.}
}
\value{
Data.frame labeled with models from Mellor et al
}
\description{
Placeholder function for extracting the outputs from the Norovirus paper
}
