% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{get_plot_cdf_by_weekday}
\alias{get_plot_cdf_by_weekday}
\title{Get a plot of the cdf colored by weekday}
\usage{
get_plot_cdf_by_weekday(delay_dfs)
}
\value{
ggplot object
}
\description{
Get a plot of the cdf colored by weekday
}
