% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_noro.R
\name{get_plot_distrib_delays}
\alias{get_plot_distrib_delays}
\title{Get plot of the distribution of delays by weekday}
\usage{
get_plot_distrib_delays(delay_dfs, n_history_delay_filter = 28)
}
\arguments{
\item{delay_dfs}{Data.frame of delay estimates both separately by weekday
and jointly, indexed by the weekday of the reference date.}

\item{n_history_delay_filter}{Integer indicating which \code{n_history_delay}
permutation to use. Default is \code{28}.}
}
\value{
ggplot object
}
\description{
Get plot of the distribution of delays by weekday
}
