% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_hub_validation.R
\name{get_plot_rel_wis_by_age_group}
\alias{get_plot_rel_wis_by_age_group}
\title{Get a plot of relative WIS by age group}
\usage{
get_plot_rel_wis_by_age_group(scores_by_age_group)
}
\arguments{
\item{scores_by_age_group}{Dataframe of the summarised scores by age group
and model}
}
\value{
ggplot object
}
\description{
Get a plot of relative WIS by age group
}
