% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{get_plot_rel_wis_over_time_mp}
\alias{get_plot_rel_wis_over_time_mp}
\title{Get a line plot of the relative wis over time for each model permutation}
\usage{
get_plot_rel_wis_over_time_mp(scores, strata)
}
\arguments{
\item{scores}{Data.frame of model permutations for all nowcast dates,
horizons, and age groups.}

\item{strata}{Character string indicating whether to summarise across the
different age strata (\code{"age groups"}) or across the nation as a whole
(\code{"national"}). Default is \code{"age groups"}.}
}
\value{
ggplot object
}
\description{
Get a line plot of the relative wis over time for each model permutation
}
