% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures_model_permutations.R
\name{make_fig_model_perms}
\alias{make_fig_model_perms}
\title{Title}
\usage{
make_fig_model_perms(
  plot_nowcasts_over_time_mp,
  bar_chart_wis_by_mp,
  rel_wis_over_time_mp,
  bar_chart_coverage_mp,
  rel_wis_by_horizon_mp,
  rel_decomp_wis_by_age_group,
  fig_file_name,
  fig_file_dir = file.path("output", "figs"),
  save = TRUE
)
}
\arguments{
\item{plot_nowcasts_over_time_mp}{A}

\item{bar_chart_wis_by_mp}{B}

\item{rel_wis_over_time_mp}{C}

\item{bar_chart_coverage_mp}{D}

\item{rel_wis_by_horizon_mp}{E}

\item{rel_decomp_wis_by_age_group}{F}

\item{fig_file_name}{Character string indicating name of the figure to be
saved as the file name}

\item{fig_file_dir}{Path to save figure. Default is
\code{file.path("output", "figs")}.}

\item{save}{Boolean indicating whether or not to save the figure to disk.
Default is \code{TRUE}.}
}
\value{
ggplot object as a gridded panel
}
\description{
Title
}
